+++ src/main/java/com/ogc/standard/bo/IHLOrderBO.java
@@ -1,5 +1,6 @@
 package com.ogc.standard.bo;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import com.ogc.standard.bo.base.IPaginableBO;
@@ -10,7 +11,7 @@ import com.ogc.standard.enums.EHLOrderStatus;
 
 public interface IHLOrderBO extends IPaginableBO<HLOrder> {
 
-    String applyOrder(Account account, Jour jour, Long applyAmount,
+    String applyOrder(Account account, Jour jour, BigDecimal applyAmount,
             String applyUser, String applyNote);
 
     void approveOrder(HLOrder order, EHLOrderStatus status, String approveUser,
